Config: Temporarily block ceb.wikipedia.org
Recently ceb wiki protected all of tehir templates which made
change-prop accumulate an enormous backlog for that wiki.

Sadly, the if-modified-since optimisation didn't work as
it clashed with the no-content-change optimisation (already fixed
for the future for this case).

To quickly reduce the backlog let's just ignore the ceb.wiki for some
time, it's almost entirely bot generated, also these reparses are no-op.

Change-Id: I31dde948f4c49251e3d594f8d2ea29778c5cd236