[Config] Block Module namespace rerenders.
The 828 Module namespace is used to store the Scribunto
Lua scripts which are counted as a dependency on any page
that accessed them during parsing creating enormous update chains.

The output of these scripts is dynamic, so edot of the script
shouldn't really provoke a rerender of all the pages using it -
it shouldn't be expected due to the dynamic nature of this.

Change-Id: Ibcee40c1768a71b1070afc591f89a1a745e51124