new rule: purge everything in resource-purge topic
We've realised that some stuff hasn't been being purged that should
be. We purge things outside of the explicit varnish purging rule and
this was missed in our migration to k8s. We're going to publish all
the purges that we should have been seeing to this new topic and then
in scb (where we can still access multicast adresses) we will read and
purge these events. We remove purge_varnish also because we're already
generating those events in k8s.

Change-Id: I183cfa6b9868536d86da567e56f092f32352a328